requirements.txt
amqp
annotated-types
anyio
async-timeout
bcrypt
billiard
celery
cffi
click
click-didyoumean
click-plugins
click-repl
colorama
cryptography
dnspython
ecdsa
exceptiongroup
fastapi
greenlet
h11
idna
jinja2
jose
kafka-python
kombu
MarkupSafe
mysqlclient
passlib
prometheus-client
prometheus-fastapi-instrumentator
prompt-toolkit
pyasn1
pycparser
pydantic
pydantic-core
pymongo
PyMySQL
python-dateutil
python-dotenv
python-jose
python-multipart
redis
rsa
six
sniffio
SQLAlchemy
starlette
typing-extensions
tzdata
uvicorn
vine
wcwidth
